+++ CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-05-25
+
+### Added
+
+- Widget component (#21).
+- Table component (#20).
+- Modal component (#19).
+- Form item component (#18).
+- Input Textarea (#17).
+
+### Fixed
+
+- Add outline focus to input and select (#16).
+
+## [0.1.3] - 2026-05-22
+
+### Fixed
+
+- Remove file class list for non file input (#14).
+
 ## [0.1.0] - 2026-05-21
 
 ### Added