+++ docs/changelog.md
@@ -10,6 +10,26 @@ import { version } from '../package.json'
 
 Current version: {{ version }}
 
+## v0.2.0 - 2026-05-25
+
+### Added
+
+- Widget component.
+- Table component.
+- Modal component.
+- Form item component.
+- Input Textarea.
+
+### Fixed
+
+- Add outline focus to input and select.
+
+## v0.1.3 - 2026-05-22
+
+### Fixed
+
+- Remove file class list for non file input.
+
 ## v0.1.0 - 2026-05-21
 
 ### Added